fibinocci/requirements.txt
Flask==0.8
gunicorn==19.4.5
requests==0.11.1
PySocks==1.6.7
beautifulsoup4==4.1.3

